+++ docs/source/notebooks/sampling_paths/agent.py
@@ -1,4 +1,3 @@
-from collections.abc import Callable
 from typing import TYPE_CHECKING, Any
 
 import equinox as eqx
@@ -17,7 +16,8 @@ from jaxtyping import (
 
 from differt.scene import TriangleScene
 
-from .metrics import reward
+from .generators import random_scene
+from .metrics import accuracy, hit_rate, reward_fn
 from .model import Model
 
 if TYPE_CHECKING:
@@ -29,89 +29,38 @@ else:
 def flow_matching_loss(
     model: Model,
     scene: TriangleScene,
-    espilon: Float[Array, ""],
     key: PRNGKeyArray,
-    reward_fn: Callable[[Int[Array, "order"], TriangleScene], Float[Array, ""]]
-    | None = None,
-) -> tuple[Float[Array, ""], tuple[Int[Array, "order"], Float[Array, ""]]]:
+) -> Float[Array, " "]:
     """
     Compute the flow matching loss for a given model and scene.
     """
-    partial_path_candidate = -jnp.ones(model.order, dtype=int)
-    last_object = jnp.array(-1)
-    if reward_fn is None:
-        reward_fn = reward
-    parent_flow_key, key = jr.split(key)
-    parent_flows = model.flow(
-        scene,
-        partial_path_candidate,
-        last_object,
-        key=parent_flow_key,
-    )
-
-    # Sums of flow mismatch
-    flow_mismatch = jnp.array(0.0)
-
-    for i, key in enumerate(jr.split(key, model.order)):
-        edge_flow_key, action_key, greedy_key = jr.split(key, 3)
-
-        action = jnp.where(
-            jr.uniform(greedy_key) < espilon,
-            jr.choice(
-                action_key,
-                parent_flows.size,
-                p=(parent_flows > 0).astype(parent_flows.dtype),
-            ),
-            jr.categorical(action_key, logits=jnp.log(parent_flows)),
-        )
-        partial_path_candidate = partial_path_candidate.at[i].set(action)
-        last_object = action
-
-        if i == model.order - 1:
-            R = reward_fn(partial_path_candidate, scene)
-            edge_flows = jnp.zeros_like(parent_flows)
-        else:
-            R = 0.0
-            edge_flows = model.flow(
-                scene,
-                partial_path_candidate,
-                last_object,
-                key=edge_flow_key,
-            )
-
-        flow_mismatch += (parent_flows[action] - edge_flows.sum() - R) ** 2
-
-        parent_flows = edge_flows
-
-    return flow_mismatch, (partial_path_candidate, R)
+    path_candidate, flows = model(scene, inference=False, key=key)
+    parent_flows = jnp.take(flows, path_candidate, axis=0)
+    sum_edge_flows = flows.sum(axis=0)
+    sum_edge_flows = jnp.roll(sum_edge_flows, -1)
+    sum_edge_flows = sum_edge_flows.at[-1].set(reward_fn(path_candidate, scene))
+    flows_mismatch = (parent_flows - sum_edge_flows) ** 2
+    return flows_mismatch.sum()
 
 
 def loss(
     model: Model,
     scene: TriangleScene,
-    espilon: Float[ArrayLike, ""],
-    keys: Key[Array, " batch_size"],
-    reward_fn: Callable[[Int[Array, "order"], TriangleScene], Float[Array, ""]]
-    | None = None,
-) -> tuple[
-    Float[Array, " "],
-    tuple[Int[Array, "batch_size order"], Float[Array, " batch_size"]],
-]:
-    loss_values, aux = jax.vmap(
-        flow_matching_loss, in_axes=(None, None, None, 0, None)
-    )(model, scene, espilon, keys, reward_fn)
-    return loss_values.mean(), aux
+    batch_size: int,
+    key: PRNGKeyArray,
+) -> Float[Array, " "]:
+    loss_values = jax.vmap(flow_matching_loss, in_axes=(None, None, 0))(
+        model, scene, jr.split(key, batch_size)
+    )
+    return loss_values.mean()
 
 
 class Agent(eqx.Module):
     """
-    Agent that learns to sample path candidates on triangle scenes.
+    Agent that trains a model to learn to sample path candidates on triangle scenes.
+
     It uses flow matching to learn a flow model that estimates the flow of paths
     through the scene, and uses this flow model to sample path candidates.
-
-    Additionally, it uses a memory to store past experiences of sampled path candidates
-    and their rewards, which can be used to quickly compute averaged rewards and train the flow model
-    on path experiences by imposing that the total flow matches the total number of valid paths.
     """
 
     # Static
@@ -124,9 +73,6 @@ class Agent(eqx.Module):
     optim: optax.GradientTransformationExtraArgs
     opt_state: optax.OptState
     steps_count: Int[Array, " "]
-    epsilon: Float[Array, ""]
-    delta_epsilon: Float[Array, ""]
-    min_epsilon: Float[Array, ""]
 
     def __init__(
         self,
@@ -139,15 +85,11 @@ class Agent(eqx.Module):
         min_epsilon: Float[ArrayLike, ""] = 0.1,
     ) -> None:
         self.batch_size = batch_size
-
         self.model = model
 
         self.optim = optax.adam(3e-5) if optim is None else optim
         self.opt_state = self.optim.init(eqx.filter(self.model, eqx.is_array))
         self.steps_count = jnp.array(0)
-        self.epsilon = jnp.array(epsilon)
-        self.delta_epsilon = jnp.array(delta_epsilon)
-        self.min_epsilon = jnp.array(min_epsilon)
 
     @eqx.filter_jit
     def train(
@@ -155,23 +97,23 @@ class Agent(eqx.Module):
         scene: TriangleScene,
         *,
         key: PRNGKeyArray,
-        reward_fn: Callable[
-            [Int[Array, "order"], TriangleScene], Float[Array, ""]
-        ]
-        | None = None,
-    ) -> tuple[Self, Float[Array, " "], Float[Array, " "]]:
+    ) -> tuple[Self, Float[Array, " "]]:
         """
         Train the model on one scene using the flow matching loss.
+
+        Args:
+            scene: The scene to train on.
+            key: The key to use for training.
+
+        Returns:
+            The updated agent and the average loss value.
         """
 
-        (loss_value, (path_candidates, rewards)), grads = (
-            eqx.filter_value_and_grad(loss, has_aux=True)(
-                self.model,
-                scene,
-                self.epsilon,
-                jr.split(key, self.batch_size),
-                reward_fn,
-            )
+        loss_value, grads = eqx.filter_value_and_grad(loss)(
+            self.model,
+            scene,
+            batch_size=self.batch_size,
+            key=key,
         )
 
         updates, opt_state = self.optim.update(
@@ -184,18 +126,52 @@ class Agent(eqx.Module):
                     agent.model,
                     agent.opt_state,
                     agent.steps_count,
-                    agent.epsilon,
                 ),
                 self,
                 (
                     eqx.apply_updates(self.model, updates),
                     opt_state,
                     self.steps_count + 1,
-                    (self.epsilon - self.delta_epsilon).clip(
-                        min=self.min_epsilon
-                    ),
                 ),
             ),
             loss_value,
-            rewards.mean(),
         )
+
+    @eqx.filter_jit
+    def evaluate(
+        self,
+        scene_keys: Key[Array, " num_scenes"],
+        *,
+        num_path_candidates: int = 10,
+        key: PRNGKeyArray,
+    ) -> tuple[Float[Array, " "], Float[Array, " "]]:
+        """
+        Evaluate the model accuracy and hit rate on a sequence of scenes.
+
+        Args:
+            scene_keys: The scene keys to generate scenes on which to evaluate the model.
+            num_path_candidates: The number of path candidates that will be
+                generated to compute the accuracy and hit rate.
+            key: The random key to be used.
+
+        Returns:
+            The average accuracy and average hit rate.
+        """
+
+        def _evaluate(
+            scene_key: PRNGKeyArray, key: PRNGKeyArray
+        ) -> tuple[Float[Array, " "], Float[Array, " "]]:
+            scene = random_scene(key=scene_key)
+            path_candidates = jax.vmap(
+                lambda key: self.model(scene, inference=True, key=key)
+            )(jr.split(key, num_path_candidates))
+            return accuracy(scene, path_candidates), hit_rate(
+                scene, path_candidates
+            )
+
+        num_scenes = scene_keys.shape[0]
+        keys = jr.split(key, num_scenes)
+
+        accuracies, hit_rates = jax.vmap(_evaluate)(scene_keys, keys)
+
+        return accuracies.mean(), jnp.nanmean(hit_rates)